build: track upstream main branch, not tag
The testnet release process involves updating Galileo's Cargo.toml for
the most recent testnet tag. While it's grand to have a 1:1
correspondence between Galileo's deps and the testnet version it handles
disbursements for, it's quite tedious to bump the deps every week. Let's
try tracking main branch, instead. Ultimately we aim to automate the
build and rollout of galileo updates, but for now, a human's attention
is still required.