build: re-add cargo lockfile
We've oscillated on whether or not to include the lockfile
in this binary application. Because of our use of git paths
for penumbra deps, we've ignored the lock file from the repo,
and let it be autogenerated on each build. This makes it hard
to reason about what's actually being used to build the binary,
though, breaking reproducibility between CI and local dev:
just now, after release of v0.70.0, the built container failed
but running galileo from my workstation succeeded.

I'm re-adding the lockfile so it's clear what the build inputs are.