chore: 🔐 check `Cargo.lock` into git
the gitignore file, prior to this commit, states:

```
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
```

this points to the rust documentation outlining lockfile practices,
here: https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html

Galileo is a binary, with a top-level `src/main.rs`. so, we should
follow that advice and remove `Cargo.lock` from the gitignore, and check
it into the repository.